Refactored server code in preparation for cf-testd
