+++ cf-serverd/server_tls.h
@@ -37,8 +37,12 @@
 bool ServerTLSInitialize();
 void ServerTLSDeInitialize();
 int ServerTLSPeek(ConnectionInfo *conn_info);
+int BasicServerTLSSessionEstablish(ServerConnectionState *conn);
 int ServerTLSSessionEstablish(ServerConnectionState *conn);
 bool BusyWithNewProtocol(EvalContext *ctx, ServerConnectionState *conn);
+int ServerSendWelcome(const ServerConnectionState *conn);
+bool ServerIdentificationDialog(ConnectionInfo *conn_info,
+                                char *username, size_t username_size);
 
 
 #endif  /* CFENGINE_SERVER_TLS_H */
